Dev

See merge request joiest/jpf!37

+++ jpf-domain/src/main/java/com/joiest/jpf/common/po/PayShopBargainRequest.java
@@ -1,5 +1,5 @@
 package com.joiest.jpf.common.po;
-
+import java.math.BigDecimal;
 import java.io.Serializable;
 import java.util.Date;
 
@@ -22,7 +22,7 @@ public class PayShopBargainRequest implements Serializable {
     /**
      * 最低豆数
      */
-    private Integer minDou;
+    private BigDecimal minDou;
 
     /**
      * 是否开启此收豆信息 0=不开启 1=开启
@@ -60,11 +60,11 @@ public class PayShopBargainRequest implements Serializable {
         this.offRate = offRate;
     }
 
-    public Integer getMinDou() {
+    public BigDecimal getMinDou() {
         return minDou;
     }
 
-    public void setMinDou(Integer minDou) {
+    public void setMinDou(BigDecimal minDou) {
         this.minDou = minDou;
     }
